requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1

python-dotenv==1.0.1
PyJWT==2.10.1
google-auth==2.25.0

requests==2.31.0

# Supabase client compatible with Python 3.12 + Pydantic v2
supabase==2.4.6
httpx==0.27.0

# Razorpay SDK
razorpay==1.3.0

# Pydantic v2 (required by Python 3.12)
pydantic==2.5.3
pydantic_core==2.14.6

python-multipart==0.0.6
Pillow==10.0.0
